# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/audit/manager-reports": {
      "get": {
        "operationId": "AuditController_managerReports",
        "parameters": [
          {
            "name": "category",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actorRole",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Cross-manager oversight feed — role-scoped exclusions applied server-side",
        "tags": [
          "audit"
        ]
      }
    },
...
  },
  "info": {
    "title": "blujet API",
    "description": "blujet airline platform — internal API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}